+++ docs/release-notes.rst
@@ -10,6 +10,8 @@ Release Notes
     * Added docs about same module name in another python redis cluster project.
     * Fix a bug when a connection was to be tracked for a node but the node either do not yet exists or
       was removed because of resharding was done in another thread. (ashishbaghudana)
+    * Fixed a bug with "CLUSTER ..." commands when a node_id argument was needed and the return type
+      was supposed to be converted to bool with bool_ok in redis._compat.
 
 
 1.2.0 (Apr 09, 2016)